system: Move the host page into the system component
While not all systemd specific code, most of the pages functionality
depends on systemd or related services. Once we support a second
init system, we can reexamine which code will be shared.

Closes #2756
