+++ pkg/shell/Makefile.am
@@ -3,7 +3,6 @@ shell_GZ = \
 	pkg/shell/shell.min.js.gz \
 	pkg/shell/bundle.min.js.gz \
 	pkg/shell/shell.min.css.gz \
-	pkg/shell/shell.min.html.gz \
 	$(NULL)
 shell_DATA = \
 	pkg/shell/index.min.html \
@@ -24,16 +23,12 @@ shelldebug_DATA = \
 	pkg/shell/index.js \
 	pkg/shell/machines.js \
 	pkg/shell/shell.css \
-	pkg/shell/shell.html \
 	pkg/shell/dbus.js \
 	pkg/shell/controls.js \
 	pkg/shell/plot.js \
 	pkg/shell/cockpit-util.js \
 	pkg/shell/cockpit-main.js \
 	pkg/shell/cockpit-plot.js \
-	pkg/shell/cockpit-server.js \
-	pkg/shell/cockpit-cpu-status.js \
-	pkg/shell/cockpit-memory-status.js \
 	pkg/shell/authorized-keys.js \
 	$(shell_SCRIPTS) \
 	$(NULL)
@@ -59,9 +54,6 @@ shell_MINFILES = \
 	pkg/shell/cockpit-util.min.js \
 	pkg/shell/cockpit-main.min.js \
 	pkg/shell/cockpit-plot.min.js \
-	pkg/shell/cockpit-server.min.js \
-	pkg/shell/cockpit-cpu-status.min.js \
-	pkg/shell/cockpit-memory-status.min.js \
 	pkg/shell/authorized-keys.min.js \
 	$(NULL)
 
@@ -101,7 +93,6 @@ CLEANFILES += \
 	pkg/shell/shell.min.js \
 	pkg/shell/bundle.min.js \
 	pkg/shell/index.min.html \
-	pkg/shell/shell.min.html \
 	$(shell_BUNDLEFILES) \
 	$(shell_GZ) \
 	$(shell_MINFILES) \
@@ -112,7 +103,6 @@ EXTRA_DIST += \
 	pkg/shell/shell.min.js \
 	pkg/shell/bundle.min.js \
 	pkg/shell/index.min.html \
-	pkg/shell/shell.min.html \
 	$(shell_DATA) \
 	$(shell_BUNDLE) \
 	$(shell_SHELL) \
