+++ pkg/shell/manifest.json
@@ -6,14 +6,5 @@
         "de": "Deutsch",
         "da": "Dansk",
         "pl": "Polski"
-    },
-
-    "menu": {
-        "_host_": {
-            "path": "system/host",
-            "label": "System",
-            "order": 0
-        }
     }
-
 }
