+++ pkg/systemd/manifest.json
@@ -3,6 +3,10 @@
     "alias": [ "server", "system" ],
 
     "menu": {
+        "host": {
+            "label": "System",
+            "order": 0
+        },
         "init": {
             "label": "Services",
             "order": 5
