+++ po/POTFILES.in
@@ -24,13 +24,9 @@ pkg/kubernetes/nulecule.js
 pkg/networkmanager/interfaces.js
 [type: gettext/glade] pkg/realmd/operation.html
 pkg/realmd/operation.js
-[type: gettext/glade] pkg/shell/shell.html
 pkg/shell/authorized-keys.js
-pkg/shell/cockpit-cpu-status.js
 pkg/shell/cockpit-main.js
-pkg/shell/cockpit-memory-status.js
 pkg/shell/cockpit-plot.js
-pkg/shell/cockpit-server.js
 pkg/shell/cockpit-util.js
 [type: gettext/glade] pkg/shell/index.html
 pkg/shell/index.js
@@ -43,6 +39,8 @@ pkg/storaged/permissions.js
 pkg/storaged/devices.js
 [type: gettext/glade] pkg/subscriptions/subscriptions.html
 pkg/subscriptions/subscriptions.js
+[type: gettext/glade] pkg/systemd/host.html
+pkg/systemd/host.js
 [type: gettext/glade] pkg/systemd/init.html
 pkg/systemd/init.js
 [type: gettext/glade] pkg/systemd/journal.html
