+++ test/check-multi-os
@@ -51,13 +51,15 @@ def add_machine(b, address):
 
 def test_spawn(b, address):
     b.call_js_func("""(function (address) {
-      window.test_result = null;
-      cockpit.spawn([ "echo", "hi" ], { host: address }).
-        done(function (data) {
-          window.test_result = data;
-        }).
-        fail(function (error) {
-          console.log(error);
+        require(["base1/cockpit"], function(cockpit) {
+            window.test_result = null;
+            cockpit.spawn([ "echo", "hi" ], { host: address }).
+                done(function (data) {
+                    window.test_result = data;
+                }).
+                fail(function (error) {
+                    console.log(error);
+                });
         });
     })""", address)
 
@@ -66,15 +68,18 @@ def test_spawn(b, address):
 
 def test_dbus(b, address):
     b.call_js_func("""(function (address) {
-      window.test_result = null;
-      var p = cockpit.dbus("org.freedesktop.DBus", { host: address }).proxy("org.freedesktop.DBus", "/");
-      p.wait(function () {
-        if (p.valid)
-          p.GetId().
-            done(function (id) {
-              window.test_result = id;
+        require(["base1/cockpit"], function(cockpit) {
+            window.test_result = null;
+            var p = cockpit.dbus("org.freedesktop.DBus", { host: address }).proxy("org.freedesktop.DBus", "/");
+            p.wait(function() {
+                if (p.valid) {
+                    p.GetId().
+                        done(function (id) {
+                            window.test_result = id;
+                        });
+                }
             });
-      });
+        });
     })""", address)
 
     b.wait_js_func("(function () { return window.test_result != null; })");
