+++ test/check-session
@@ -46,7 +46,7 @@ class TestSession(MachineCase):
         wait_session(False)
 
         # Login
-        self.login_and_go("server", href="/system/host")
+        self.login_and_go("/system/host")
         wait_session(True)
 
         # Logout
@@ -59,14 +59,14 @@ class TestSession(MachineCase):
         b.set_val("#login-password-input", "foobar")
         b.click('#login-button')
         b.expect_reload()
-        b.enter_page("server")
+        b.enter_page("/system/host")
         wait_session(True)
 
         # Kill session from the outside
         m.execute("loginctl kill-user admin")
         wait_session(False)
 
-        b.relogin("server", "admin")
+        b.relogin("/system/host", "admin")
         wait_session(True)
 
         # Kill session from the inside
