+++ test/check-system-info
@@ -42,7 +42,7 @@ class TestSystemInfo(MachineCase):
         m.execute("rm /etc/os-release")
         m.write("/etc/os-release", os_release)
 
-        self.login_and_go("server", href="/system/host")
+        self.login_and_go("/system/host")
         b.wait_text('#system_information_os_text',
                     "Foobar Adventure Linux Server 2.0 (Day of Doom)")
 
@@ -82,7 +82,7 @@ class TestSystemInfo(MachineCase):
         # make sure system is on expected timezone EEST
         m.execute("LC_ALL=C timedatectl set-timezone Europe/Helsinki")
 
-        self.login_and_go("server", href="/system/host")
+        self.login_and_go("/system/host")
 
         # Make sure the date loads
         b.wait_text_not("#system_information_systime_button", "")
